load theme changer sooner
on the homepage especially, which takes a moment to load, there's a
moment when the page has partially loaded but the theme changer hasn't
been activated yet, where if one is using the dark theme one is
subjected to blindly bright light. loading the theme changer sooner
immediately makes the page dark.